refactor(cli): extract @kubb/middleware-logger and drop @clack/prompts
Moves all log-related logic out of @kubb/cli into a new @kubb/middleware-logger
package (mirroring the @kubb/middleware-barrel shape). The package exports
createLogger(options) and a default middlewareLogger instance: one consola-backed
logger that inlines GitHub Actions workflow command annotations when the runner
is detected, so groups, warnings, errors, and notices show up in the CI
annotations panel without a separate decorator pass.

@kubb/cli now imports middlewareLogger from the new package and is otherwise
just a consumer. @clack/prompts is fully removed; the init wizard, agent runner,
and generate watcher use consola for prompts and live output.

@kubb/core's createReporter no-flush branch is patched to discard the generic T
return value so the resulting Reporter.report matches its void contract.
