+++ .changeset/cli-unified-consola-logger.md
@@ -1,7 +1,13 @@
 ---
+'@kubb/middleware-logger': minor
 '@kubb/cli': minor
+'@kubb/core': patch
 ---
 
-Collapse the `clack` / `github-actions` / `plain` CLI loggers into a single [consola](https://github.com/unjs/consola)-backed logger. One set of `KubbHooks` handlers now covers every environment, with an optional decorator that mirrors warnings, errors, notices, and section boundaries as GitHub Actions workflow commands (`::group::`, `::warning::`, `::error::`, `::notice::`) when the runner is detected.
+Extract the CLI logger into a new `@kubb/middleware-logger` package. A single consola-backed logger now handles every output environment, and when the runtime is GitHub Actions the same handlers inline workflow command annotations (`::group::`, `::endgroup::`, `::warning::`, `::error::`, `::notice::`) so the CI annotations panel highlights problems and sections collapse per config, plugin, and hook.
 
-The reporters (`cli`, `json`, `file`) are unchanged. CLI output stays familiar — colored on TTY, plain in CI, annotated under GitHub Actions — but the in-place spinners, progress bars, and per-hook task logs that came from `@clack/prompts` are replaced by consola's line-based output. Hook subprocess output now streams through `consola.log` rather than into a collapsible block.
+The package exports `createLogger(options)` as the factory and a default `middlewareLogger` instance, plus the diagnostic and formatting helpers (`formatDiagnostic`, `diagnosticSymbol`, `diagnosticHeadline`, `diagnosticDetails`, `formatMessage`, `formatCommandWithArgs`, `createHookTimer`) and the `HookSinkFactory` / `HookSinkOptions` types so other hosts can reuse them.
+
+`@clack/prompts` is fully removed from `@kubb/cli`. The init wizard, agent runner, and generate watcher now go through consola for prompts and live output.
+
+`@kubb/core` patches a type bug in `createReporter<T>`'s no-`flush` branch so the returned `Reporter.report` correctly discards the `T` value rather than mistyping the callback.
