+++ packages/middleware-logger/package.json
@@ -1,0 +1,69 @@
+{
+  "name": "@kubb/middleware-logger",
+  "version": "5.0.0-beta.37",
+  "description": "Logger middleware for Kubb. One consola-backed logger that inlines GitHub Actions workflow command annotations when the runtime is detected, for use by the CLI and other hosts.",
+  "keywords": [
+    "codegen",
+    "consola",
+    "github-actions",
+    "kubb",
+    "logger",
+    "middleware"
+  ],
+  "license": "MIT",
+  "author": "stijnvanhulle",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/kubb-labs/kubb.git",
+    "directory": "packages/middleware-logger"
+  },
+  "files": [
+    "src",
+    "dist",
+    "*.d.ts",
+    "*.d.cts",
+    "!/**/**.test.**",
+    "!/**/__tests__/**",
+    "!/**/__snapshots__/**"
+  ],
+  "type": "module",
+  "sideEffects": false,
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    },
+    "./package.json": "./package.json"
+  },
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org/"
+  },
+  "scripts": {
+    "build": "tsdown",
+    "clean": "npx rimraf ./dist",
+    "lint": "oxlint .",
+    "lint:fix": "oxlint --fix .",
+    "release": "pnpm publish --no-git-check",
+    "release:stage": "pnpm stage publish --no-git-check",
+    "start": "tsdown --watch",
+    "test": "vitest --passWithNoTests",
+    "typecheck": "tsc -p ./tsconfig.json --noEmit --emitDeclarationOnly false"
+  },
+  "dependencies": {
+    "consola": "^3.4.2"
+  },
+  "devDependencies": {
+    "@internals/utils": "workspace:*",
+    "@kubb/core": "workspace:*"
+  },
+  "peerDependencies": {
+    "@kubb/core": "workspace:*"
+  },
+  "engines": {
+    "node": ">=22"
+  }
+}
