+++ packages/middleware-logger/src/diagnostics.ts
