+++ packages/middleware-logger/tsdown.config.ts
@@ -1,0 +1,35 @@
+import { defineConfig, type UserConfig } from 'tsdown'
+
+const entry = {
+  index: 'src/index.ts',
+}
+
+const shared: Partial<UserConfig> = {
+  platform: 'node',
+  sourcemap: true,
+  shims: true,
+  exports: true,
+  deps: {
+    neverBundle: [/^@kubb\//],
+    alwaysBundle: [/@internals/],
+  },
+  fixedExtension: false,
+  outputOptions: {
+    keepNames: true,
+  },
+}
+
+export default defineConfig([
+  {
+    entry,
+    format: 'esm',
+    dts: true,
+    ...shared,
+  },
+  {
+    entry,
+    format: 'cjs',
+    dts: false,
+    ...shared,
+  },
+])
