+++ packages/middleware-logger/vitest.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    dir: './src',
+  },
+  resolve: {
+    tsconfigPaths: true,
+  },
+})
